Fixed AnswerCall Attribute Of Ring BXML Verb
Changed the answerCall field of the Ring class from a boolean primitive to a Boolean wrapper for a default 'null' value if not explicitly set. Otherwise, the boolean primitive would default to 'true' and the attribute would be serialized as such, even if the value wasn't set. Also added a test to assert values that aren't set aren't serialized into the BXML.